+++ packages/scripts/src/monitoring-dvm2.0/MonitorCommon.ts
