+++ packages/scripts/src/monitoring-dvm2.0/MonitorEmergency.ts
@@ -2,7 +2,7 @@ import { Logger } from "@uma/financial-templates-lib";
 import { EmergencyProposerEthers } from "@uma/contracts-node";
 import { logEmergencyProposal } from "./MonitorLogger";
 import { getContractInstanceByUrl } from "../utils/contracts";
-import type { MonitoringParams } from "./common";
+import type { MonitoringParams } from "./MonitorCommon";
 
 export async function monitorEmergency(logger: typeof Logger, params: MonitoringParams): Promise<void> {
   const emergencyProposer = await getContractInstanceByUrl<EmergencyProposerEthers>(
