+++ packages/scripts/src/monitoring-dvm2.0/MonitorRolled.ts
@@ -1,9 +1,9 @@
 import { Logger } from "@uma/financial-templates-lib";
 import { VotingV2Ethers } from "@uma/contracts-node";
-import { checkEndBlockVotingRound, getRequestId } from "./common";
+import { checkEndBlockVotingRound, getRequestId } from "./MonitorCommon";
 import { logRolled } from "./MonitorLogger";
 import { getContractInstanceByUrl } from "../utils/contracts";
-import type { MonitoringParams } from "./common";
+import type { MonitoringParams } from "./MonitorCommon";
 
 export async function monitorRolled(logger: typeof Logger, params: MonitoringParams): Promise<void> {
   const votingV2 = await getContractInstanceByUrl<VotingV2Ethers>("VotingV2", params.jsonRpcUrl);
