+++ packages/scripts/src/monitoring-dvm2.0/index.ts
@@ -1,5 +1,5 @@
 import { delay, Logger } from "@uma/financial-templates-lib";
-import { initCommonEnvVars, updateBlockRange, startupLogLevel } from "./common";
+import { initCommonEnvVars, updateBlockRange, startupLogLevel } from "./MonitorCommon";
 import { monitorUnstakes } from "./MonitorUnstakes";
 import { monitorStakes } from "./MonitorStakes";
 import { monitorGovernance } from "./MonitorGovernance";
